requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-decouple==3.8
pyjwt[crypto]==2.8.0
requests==2.31.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
httpx==0.25.2
pandas==2.1.4
numpy==1.24.3
scipy==1.11.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0
prometheus-client==0.19.0
# Observability
# opentelemetry-api==1.21.0
# opentelemetry-sdk==1.21.0
# opentelemetry-auto-instrumentation==0.42b0
# opentelemetry-exporter-jaeger==1.21.0
structlog==23.2.0
